+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 # Upcoming
 
 ### ✅ Added
+- Added markdown support (enabled by default)
+
+# [4.51.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.51.0)
+_March 26, 2024_
+
+### ✅ Added
 - Role value in the user display info
 
 ### 🐞 Fixed